backend/pyproject.toml
[project]
name = "exoplanets-rag-backend"
version = "1.0.0"
description = "FastAPI backend for Exoplanets Retrieval-Augmented Generation"
authors = [{name = "Fernando Dilland"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "aiomysql==0.2.0",
    "sqlalchemy==2.0.35",
    "asyncmy==0.2.9",
    "argon2-cffi==23.1.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.12",
    "python-dotenv==1.0.1",
    "httpx==0.27.2",
    "aioboto3==13.2.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
